Added more specified labels for generalized y value, got rid of default long description, and created submodules for more organized code.
