+++ pyharp/__init__.py
@@ -1,1 +1,3 @@
 from .core import *
+from .labels import *
+from .media import *
