+++ pyharp/media/__init__.py
@@ -1,0 +1,2 @@
+from .audio import *
+from .midi import *
